// src/generated/model/saksStatusResponseStatus.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type SaksStatusResponseStatus = typeof SaksStatusResponseStatus[keyof typeof SaksStatusResponseStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SaksStatusResponseStatus = {
    UNDER_BEHANDLING: "UNDER_BEHANDLING",
    IKKE_INNSYN: "IKKE_INNSYN",
    FERDIGBEHANDLET: "FERDIGBEHANDLET",
    BEHANDLES_IKKE: "BEHANDLES_IKKE",
    FEILREGISTRERT: "FEILREGISTRERT",
} as const;